Moved wlauto.core.config.core into wlauto.core.configuration
