--- wlauto/core/config/__init__.py
@@ -1,2 +1,0 @@
-from wlauto.core.config.core import settings, ConfigurationPoint, PluginConfiguration
-from wlauto.core.config.core import merge_config_values, WA_CONFIGURATION
