+++ wlauto/core/execution.py
@@ -51,7 +51,6 @@ import wlauto.core.signal as signal
 from wlauto.core import instrumentation
 from wlauto.core.config.core import settings
 from wlauto.core.plugin import Artifact
-from wlauto.core.configuration import RunConfiguration
 from wlauto.core import pluginloader
 from wlauto.core.resolver import ResourceResolver
 from wlauto.core.result import ResultManager, IterationResult, RunResult
